+++ src/cpp/tardigrade_hydra.cpp
@@ -1932,6 +1932,8 @@ namespace tardigradeHydra{
 
         resetNumGrad( );
 
+        setRankDeficientError( false );
+
         try{
 
             solveNonLinearProblem( );
@@ -1951,6 +1953,8 @@ namespace tardigradeHydra{
 
             resetIterations( );
             updateUnknownVector( _initialX );
+
+            setUseLevenbergMarquardt( false );
 
             try{