+++ src/cpp/tardigrade_hydraMicromorphicDruckerPragerPlasticity.cpp
@@ -8032,6 +8032,8 @@ namespace tardigradeHydra{
 
             floatVector macNegMicroGradientGamma( 3, 0 );
 
+            floatVector consistencyConditionModuli( 5, *getConsistencyConditionModulus( ) );
+
             if ( *useWeakenedMacaulay( ) ){
 
                 macroMac  = weakMac( *macroYield, *getWeakenedMacaulayParameter( ) );
@@ -8069,16 +8071,16 @@ namespace tardigradeHydra{
             }
 
             // Set the terms associated with the yield surface
-            ( *residual.value )[ 0 ] = macroMac + ( *plasticMultipliers )[ 0 ] * ( *macroYield ) + ( *getPlasticMultiplierBarrierModulus( ) ) * macNegMacroGamma;
+            ( *residual.value )[ 0 ] = macroMac + consistencyConditionModuli[ 0 ] * std::fabs( ( *plasticMultipliers )[ 0 ] * ( *macroYield ) ) + ( *getPlasticMultiplierBarrierModulus( ) ) * macNegMacroGamma;
 
-            ( *residual.value )[ 1 ] = microMac + ( *plasticMultipliers )[ 1 ] * ( *microYield ) + ( *getPlasticMultiplierBarrierModulus( ) ) * macNegMicroGamma;
+            ( *residual.value )[ 1 ] = microMac + consistencyConditionModuli[ 1 ] * std::fabs( ( *plasticMultipliers )[ 1 ] * ( *microYield ) ) + ( *getPlasticMultiplierBarrierModulus( ) ) * macNegMicroGamma;
 
             for ( auto y = microGradientYield->begin( ); y != microGradientYield->end( ); y++ ){
 
                 unsigned int index = ( unsigned int )( y - microGradientYield->begin( ) );
 
                 ( *residual.value )[ index + 2 ]
-                    = microGradientMac[ index ] + ( *plasticMultipliers )[ index + 2 ] * ( *y ) + ( *getPlasticMultiplierBarrierModulus( ) ) * macNegMicroGradientGamma[ index ];
+                    = microGradientMac[ index ] + consistencyConditionModuli[ index + 2 ] * std::fabs( ( *plasticMultipliers )[ index + 2 ] * ( *y ) ) + ( *getPlasticMultiplierBarrierModulus( ) ) * macNegMicroGradientGamma[ index ];
 
             }
 
@@ -8161,6 +8163,16 @@ namespace tardigradeHydra{
             auto jacobian = get_setDataStorage_stateVariableJacobians( );
             jacobian.zero( numISVs * numUnknowns );
 
+            floatVector signs( 5, 0 );
+
+            signs[ 0 ] = sgn( ( *plasticMultipliers )[ 0 ] * ( *macroYield ) );
+            signs[ 1 ] = sgn( ( *plasticMultipliers )[ 1 ] * ( *microYield ) );
+            for ( unsigned int i = 0; i < ( numPlasticMultipliers - 2 ); i++ ){
+                signs[ i + 2 ] = sgn( ( *plasticMultipliers )[ i + 2 ] * ( *microGradientYield )[ i ] );
+            }
+
+            floatVector consistencyConditionModuli( 5, *getConsistencyConditionModulus( ) );
+
             // Stress Jacobians
             floatType dMacroMacdx, dMicroMacdx;
             floatType ndMacroMacdx, ndMicroMacdx;
@@ -8230,9 +8242,9 @@ namespace tardigradeHydra{
             unsigned int offset = numSecondOrderTensor;
             for ( unsigned int j = 0; j < numSecondOrderTensor; j++ ){
 
-                ( *jacobian.value )[ numUnknowns * 0 + j ] = ( dMacroMacdx  + ( *plasticMultipliers )[ 0 ] ) * ( *dMacroYielddStress )[ j ];
+                ( *jacobian.value )[ numUnknowns * 0 + j ] = ( dMacroMacdx  + consistencyConditionModuli[ 0 ] * signs[ 0 ]  * ( *plasticMultipliers )[ 0 ] ) * ( *dMacroYielddStress )[ j ];
 
-                ( *jacobian.value )[ numUnknowns * 1 + j + offset ] = ( dMicroMacdx + ( *plasticMultipliers )[ 1 ] ) * ( *dMicroYielddStress )[ j ];
+                ( *jacobian.value )[ numUnknowns * 1 + j + offset ] = ( dMicroMacdx + consistencyConditionModuli[ 1 ] * signs[ 1 ] * ( *plasticMultipliers )[ 1 ] ) * ( *dMicroYielddStress )[ j ];
 
             }
 
@@ -8241,7 +8253,7 @@ namespace tardigradeHydra{
 
                 for ( unsigned int j = 0; j < numThirdOrderTensor; j++ ){
 
-                    ( *jacobian.value )[ numUnknowns * ( i + 2 ) + j + offset ] = ( dMicroGradientMacdx[ i ] + ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddStress )[ numThirdOrderTensor * i + j ];
+                    ( *jacobian.value )[ numUnknowns * ( i + 2 ) + j + offset ] = ( dMicroGradientMacdx[ i ] + consistencyConditionModuli[ i + 2 ] * signs[ i + 2 ] * ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddStress )[ numThirdOrderTensor * i + j ];
 
                 }
 
@@ -8251,9 +8263,9 @@ namespace tardigradeHydra{
             offset = 2 * numSecondOrderTensor + numThirdOrderTensor;
             for ( unsigned int j = 0; j < ( numConfigurations - 1 ) * numSecondOrderTensor; j++ ){
 
-                ( *jacobian.value )[ numUnknowns * 0 + j + offset ] = ( dMacroMacdx + ( *plasticMultipliers )[ 0 ] ) * ( *dMacroYielddFn )[ j ];
+                ( *jacobian.value )[ numUnknowns * 0 + j + offset ] = ( dMacroMacdx + consistencyConditionModuli[ 0 ] * signs[ 0 ] * ( *plasticMultipliers )[ 0 ] ) * ( *dMacroYielddFn )[ j ];
 
-                ( *jacobian.value )[ numUnknowns * 1 + j + offset ] = ( dMicroMacdx + ( *plasticMultipliers )[ 1 ] ) * ( *dMicroYielddFn )[ j ];
+                ( *jacobian.value )[ numUnknowns * 1 + j + offset ] = ( dMicroMacdx + consistencyConditionModuli[ 1 ] * signs[ 1 ] * ( *plasticMultipliers )[ 1 ] ) * ( *dMicroYielddFn )[ j ];
 
             } 
 
@@ -8261,7 +8273,7 @@ namespace tardigradeHydra{
 
                 for ( unsigned int j = 0; j < ( numConfigurations - 1 ) * numSecondOrderTensor; j++ ){
 
-                    ( *jacobian.value )[ numUnknowns * ( i + 2 ) + j + offset ] = ( dMicroGradientMacdx[ i ] + ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddFn )[ ( numConfigurations - 1 ) * numSecondOrderTensor * i + j ];
+                    ( *jacobian.value )[ numUnknowns * ( i + 2 ) + j + offset ] = ( dMicroGradientMacdx[ i ] + consistencyConditionModuli[ i + 2 ] * signs[ i + 2 ] * ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddFn )[ ( numConfigurations - 1 ) * numSecondOrderTensor * i + j ];
 
                 }
 
@@ -8273,7 +8285,7 @@ namespace tardigradeHydra{
 
                 for ( unsigned int j = 0; j < ( numConfigurations - 1 ) * numSecondOrderTensor; j++ ){
 
-                    ( *jacobian.value )[ numUnknowns * ( i + 2 ) + j + offset ] = ( dMicroGradientMacdx[ i ] + ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddChin )[ ( numConfigurations - 1 ) * numSecondOrderTensor * i + j ];
+                    ( *jacobian.value )[ numUnknowns * ( i + 2 ) + j + offset ] = ( dMicroGradientMacdx[ i ] + consistencyConditionModuli[ i + 2 ] * signs[ i + 2 ] * ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddChin )[ ( numConfigurations - 1 ) * numSecondOrderTensor * i + j ];
 
                 }
 
@@ -8283,21 +8295,21 @@ namespace tardigradeHydra{
             // State Variable Jacobians
             offset = numConfigurations * ( 2 * numSecondOrderTensor + numThirdOrderTensor );
 
-            ( *jacobian.value )[ numUnknowns * 0 + offset + 0 ] += *macroYield - ( *getPlasticMultiplierBarrierModulus( ) ) * dMacNegMacroGammadGamma;
+            ( *jacobian.value )[ numUnknowns * 0 + offset + 0 ] += consistencyConditionModuli[ 0 ] * signs[ 0 ] * ( *macroYield ) - ( *getPlasticMultiplierBarrierModulus( ) ) * dMacNegMacroGammadGamma;
 
-            ( *jacobian.value )[ numUnknowns * 1 + offset + 1 ] += *microYield - ( *getPlasticMultiplierBarrierModulus( ) ) * dMacNegMicroGammadGamma;
+            ( *jacobian.value )[ numUnknowns * 1 + offset + 1 ] += consistencyConditionModuli[ 1 ] * signs[ 1 ] * ( *microYield ) - ( *getPlasticMultiplierBarrierModulus( ) ) * dMacNegMicroGammadGamma;
 
             for ( unsigned int i = 0; i < dim; i++ ){
 
-                ( *jacobian.value )[ numUnknowns * ( i + 2 ) + offset + i + 2 ] += ( *microGradientYield )[ i ] - ( *getPlasticMultiplierBarrierModulus( ) ) * dMacNegMicroGradientGammadGamma[ i ];
+                ( *jacobian.value )[ numUnknowns * ( i + 2 ) + offset + i + 2 ] += consistencyConditionModuli[ i + 2 ] * signs[ i + 2 ] * ( *microGradientYield )[ i ] - ( *getPlasticMultiplierBarrierModulus( ) ) * dMacNegMicroGradientGammadGamma[ i ];
 
             }
 
             for ( unsigned int j = 0; j < ( numPlasticMultipliers + numPlasticStrainLikeISVs ); j++ ){
 
-                ( *jacobian.value )[ numUnknowns * 0 + j + offset ] += ( dMacroMacdx + ( *plasticMultipliers )[ 0 ] ) * ( *dMacroYielddStateVariables )[ j ];
+                ( *jacobian.value )[ numUnknowns * 0 + j + offset ] += ( dMacroMacdx + consistencyConditionModuli[ 0 ] * signs[ 0 ] * ( *plasticMultipliers )[ 0 ] ) * ( *dMacroYielddStateVariables )[ j ];
 
-                ( *jacobian.value )[ numUnknowns * 1 + j + offset ] += ( dMicroMacdx + ( *plasticMultipliers )[ 1 ] ) * ( *dMicroYielddStateVariables )[ j ];
+                ( *jacobian.value )[ numUnknowns * 1 + j + offset ] += ( dMicroMacdx + consistencyConditionModuli[ 1 ] * signs[ 1 ] * ( *plasticMultipliers )[ 1 ] ) * ( *dMicroYielddStateVariables )[ j ];
 
             }
 
@@ -8305,7 +8317,7 @@ namespace tardigradeHydra{
 
                 for ( unsigned int j = 0; j < ( numPlasticMultipliers + numPlasticStrainLikeISVs ); j++ ){
 
-                    ( *jacobian.value )[ numUnknowns * ( i + 2 ) + j + offset ] += ( dMicroGradientMacdx[ i ] + ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddStateVariables )[ numISVs * i + j ];
+                    ( *jacobian.value )[ numUnknowns * ( i + 2 ) + j + offset ] += ( dMicroGradientMacdx[ i ] + consistencyConditionModuli[ i + 2 ] * signs[ i + 2 ] * ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddStateVariables )[ numISVs * i + j ];
 
                 }
 
@@ -8373,6 +8385,16 @@ namespace tardigradeHydra{
             auto dRdD = get_setDataStorage_dStateVariableResidualsdD( );
             dRdD.zero( get_plasticStateVariables( )->size( ) * numConfigurationUnknowns );
 
+            floatVector signs( 5, 0 );
+
+            signs[ 0 ] = sgn( ( *plasticMultipliers )[ 0 ] * ( *macroYield ) );
+            signs[ 1 ] = sgn( ( *plasticMultipliers )[ 1 ] * ( *microYield ) );
+            for ( unsigned int i = 0; i < ( numPlasticMultipliers - 2 ); i++ ){
+                signs[ i + 2 ] = sgn( ( *plasticMultipliers )[ i + 2 ] * ( *microGradientYield )[ i ] );
+            }
+
+            floatVector consistencyConditionModuli( 5, *getConsistencyConditionModulus( ) );
+
             // Deformation gradient jacobians
             floatType dMacroMacdx, dMicroMacdx;
             floatType ndMacroMacdx, ndMicroMacdx;
@@ -8423,9 +8445,9 @@ namespace tardigradeHydra{
             unsigned int offset = 0;
             for ( unsigned int j = 0; j < numSecondOrderTensor; j++ ){
 
-                ( *dRdD.value )[ numConfigurationUnknowns * 0 + j + offset ] = ( dMacroMacdx + ( *plasticMultipliers )[ 0 ] ) * ( *dMacroYielddF )[ j ];
+                ( *dRdD.value )[ numConfigurationUnknowns * 0 + j + offset ] = ( dMacroMacdx + consistencyConditionModuli[ 0 ] * signs[ 0 ] * ( *plasticMultipliers )[ 0 ] ) * ( *dMacroYielddF )[ j ];
 
-                ( *dRdD.value )[ numConfigurationUnknowns * 1 + j + offset ] = ( dMicroMacdx + ( *plasticMultipliers )[ 1 ] ) * ( *dMicroYielddF )[ j ];
+                ( *dRdD.value )[ numConfigurationUnknowns * 1 + j + offset ] = ( dMicroMacdx + consistencyConditionModuli[ 1 ] * signs[ 1 ] * ( *plasticMultipliers )[ 1 ] ) * ( *dMicroYielddF )[ j ];
 
             } 
 
@@ -8433,7 +8455,7 @@ namespace tardigradeHydra{
 
                 for ( unsigned int j = 0; j < numSecondOrderTensor; j++ ){
 
-                    ( *dRdD.value )[ numConfigurationUnknowns * ( i + 2 ) + j + offset ] = ( dMicroGradientMacdx[ i ] + ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddF )[ numSecondOrderTensor * i + j ];
+                    ( *dRdD.value )[ numConfigurationUnknowns * ( i + 2 ) + j + offset ] = ( dMicroGradientMacdx[ i ] + consistencyConditionModuli[ i + 2 ] * signs[ i + 2 ] * ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddF )[ numSecondOrderTensor * i + j ];
 
                 }
 
@@ -8445,7 +8467,7 @@ namespace tardigradeHydra{
 
                 for ( unsigned int j = 0; j < numSecondOrderTensor; j++ ){
 
-                    ( *dRdD.value )[ numConfigurationUnknowns * ( i + 2 ) + j + offset ] = ( dMicroGradientMacdx[ i ] + ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddChi )[ numSecondOrderTensor * i + j ];
+                    ( *dRdD.value )[ numConfigurationUnknowns * ( i + 2 ) + j + offset ] = ( dMicroGradientMacdx[ i ] + consistencyConditionModuli[ i + 2 ] * signs[ i + 2 ] * ( *plasticMultipliers )[ i + 2 ] ) * ( *dMicroGradientYielddChi )[ numSecondOrderTensor * i + j ];
 
                 }
 
@@ -9004,6 +9026,31 @@ namespace tardigradeHydra{
             }
 
         }
+
+//        void residual::suggestInitialIterateValues( std::vector< unsigned int >   &indices,
+//                                                    std::vector< floatType > &values ){
+//
+//            /*!
+//             * Function which is called which allows the residual to suggest initial values for given
+//             * configurations. This is called when the unknown vector is being initialized. If more than
+//             * one residual attempts to set the initial vector the last residual will override all of the others.
+//             *
+//             * After the initial iterate has been suggested, the iteration data is cleared so that the residual
+//             * starts the iteration in a clean state.
+//             * 
+//             * \param &indices: The indices of the unknown vector to set
+//             * \param &values:  The values to be set in the unknown vector
+//             */
+//
+//            // Assume no additional plastic deformation
+//            indices = std::vector< unsigned int >( getStateVariableIndices( )->begin( ),
+//                                                   getStateVariableIndices( )->begin( ) + 5 );
+//
+//            indices += ( *hydra->getNumConfigurations( ) ) * ( *hydra->getConfigurationUnknownCount( ) );
+//
+//            values  = std::vector< floatType >( 5, 0 );
+//
+//        }
 
         double residual::softLinearCohesion( const floatType &Z, const floatType &A, const floatType &c0, const floatType &rc, const floatType &cf ){
             /*!